+++ CHANGELOG.md
@@ -5,11 +5,14 @@
 - **Patch Version**: Incremented for bug fixes.
 
 ## UNRELEASED
+- Export Original Statements as docx in the Statement List
+- Allow filtering of institution tags in DpAddOrganizationList
+
+## v4.3.0 (2025-06-13)
 - Add Versioning of custom fields of segments
 - Update Elasticsearch to version 8
-
-## UNRELEASED
 - Export Original Statements as csv in the Statement List
+
 ## v4.1.0 (2025-05-21)
 ## v4.0.0 (2025-05-21)
 - Update to symfony 6.4